added error handling to makotoState reducer + context
changed reducer to start handling ErrorfulMakotoState, which is an
interface that holds a `state` and `error` field. after every action,
these fields will be fresh, meaning the error will be null if there was
no error, and vice versa. going to play with it and see how it works.
also changed partitions to hold a list of indices of tags instead of the
tag names